Modification du css, ajout d'un getter pour résumer les contenus sur la page d'acceuil ainsi que sur les pages d'administration

+++ html/view/frontend/allPostsView.php
@@ -1,11 +1,8 @@
-<?php $title = 'All Posts';
-session_start();
-?>
+<?php $title = 'All Posts'; ?>
 
-<?php ob_start();  ?>
+<?php ob_start(); ?>
 
-<?php if (!empty($_SESSION))
-{ ?> <!-- Si une session est ouvert, alors cela affiche le lien de déconnexion -->
+<?php if (!empty($_SESSION)): ?> <!-- Si une session est ouvert, alors cela affiche le lien de déconnexion -->
 
   <h1>Liste de tous les Posts</h1>
 
@@ -13,37 +10,33 @@ session_start();
     <h3>
         <?= $post->getTitle(); ?>
     </h3>
-    <div class="alignPost">
-      <p>
-          <?= $post->getCreationDateFr(); ?>
-      </p>
 
-      <p>
-          <?= $post->getContent() . "..."; ?>
-      </p>
-      <p>
-        <em><a href="index.php?action=post&amp;id=<?= $post->getId(); ?>">Commentaires...</a></em>
-      </p>
+    <div class="creationDateFr">
+        <p>Le <?= $post->getCreationDateFr(); ?> </p>
     </div>
 
-    <input class="btn btn-warning buttonCenter" type="button" value="Modifier" onclick="javascript:location.href='index.php?action=modifPost&postId=<?= $post->getId(); ?>'">
+    <div class="content">
+        <p><?= $post->getResumeContent() . "..."; ?></p>
+    </div>
+
+    <p>
+        <em><a href="index.php?action=post&amp;id=<?= $post->getId(); ?>">Commentaires...</a></em>
+    </p>
+
+    <input class="btn btn-warning buttonCenter" type="button" value="Modifier" onclick="javascript:location.href='index.php?action=updatePost&postId=<?= $post->getId(); ?>'">
 
     -
 
-    <input class="btn btn-danger buttonCenter" type="button" value="Supprimer" onclick="javascript:location.href='index.php?action=deletePost&id=<?= $post->getId(); ?>'">
+    <?php if($_SESSION['admin']): ?>
+        <input class="btn btn-danger buttonCenter" type="button" value="Supprimer" onclick="javascript:location.href='index.php?action=deletePost&id=<?= $post->getId(); ?>'">
+    <?php endif; ?>
+
+    <hr>
 
   <?php endforeach; ?>
 
-
-<?php
-  }
-  else
-  {
-    echo "vous n'etes pas connecté";
-  }
-?>
-
+<?php endif; ?>
 
 <?php $content = ob_get_clean(); ?>
 
-<?php require('/view/frontend/template.php'); ?>
+<?php require('view/frontend/template.php'); ?>
